requirements/base.txt
# Bleeding edge Django
django==1.10.4

# Configuration
django-environ==0.4.1
whitenoise==3.2.2

# Forms
django-braces==1.10.0
django-crispy-forms==1.6.1

# Models
django-model-utils==2.6
django-extended-choices==1.1.1

# Images
Pillow==3.4.2

# For user registration, either via email or social
# Well-built with regular release cycles!
django-allauth==0.29.0

# Python-PostgreSQL Database Adapter
psycopg2==2.6.2

# Unicode slugification
awesome-slugify==1.6.5

# Time zones support
pytz==2016.10

# Redis support
django-redis==4.6.0
redis>=2.10.5

# Django REST Framework
djangorestframework==3.5.3
django-filter==1.0.1
drf-nested-routers==0.11.1

# Celery
"celery[msgpack]"==4.0.2

rcssmin==1.0.6 
django-compressor==2.1


# Grapelli
django-grappelli==2.9.1


# Sentry
raven==5.32.0
